+++ tecina-api/database/migrations/2018_04_19_090937_create_food_types_translations_table.php
@@ -14,8 +14,8 @@ class CreateFoodTypesTranslationsTable extends Migration
     public function up()
     {
         Schema::create('food_types_translations', function (Blueprint $table) {
-            $table->smallInteger('id_food_type')->unsigned();
-            $table->smallInteger('id_language')->unsigned();
+            $table->tinyInteger('id_food_type')->unsigned();
+            $table->tinyInteger('id_language')->unsigned();
             $table->string('name');
 
             $table->primary(['id_food_type','id_language'],'food_type_language_primary');
