+++ tecina-api/database/migrations/2018_04_19_094427_create_categories_translations_table.php
@@ -14,8 +14,8 @@ class CreateCategoriesTranslationsTable extends Migration
     public function up()
     {
         Schema::create('categories_translations', function (Blueprint $table) {
-            $table->smallInteger('id_category')->unsigned();
-            $table->smallInteger('id_language')->unsigned();
+            $table->tinyInteger('id_category')->unsigned();
+            $table->tinyInteger('id_language')->unsigned();
             $table->string('name');
             $table->string('description')->nullable();
