+++ tecina-api/database/migrations/2018_04_19_095950_create_wine_varieties_translations_table.php
@@ -14,8 +14,8 @@ class CreateWineVarietiesTranslationsTable extends Migration
     public function up()
     {
         Schema::create('wine_variety_translations', function (Blueprint $table) {
-            $table->smallInteger('id_wine_variety')->unsigned();
-            $table->smallInteger('id_language')->unsigned();
+            $table->tinyInteger('id_wine_variety')->unsigned();
+            $table->tinyInteger('id_language')->unsigned();
             $table->string('name');
 
             $table->primary(['id_wine_variety','id_language'],'wine_variety_language_primary');
