+++ tecina-api/database/migrations/2018_04_19_182814_create_menus_wines_table.php
@@ -14,8 +14,8 @@ class CreateMenusWinesTable extends Migration
     public function up()
     {
         Schema::create('menus_wines', function (Blueprint $table) {
-            $table->smallInteger('id_wine')->unsigned();
-            $table->smallInteger('id_menu')->unsigned();
+            $table->tinyInteger('id_wine')->unsigned();
+            $table->tinyInteger('id_menu')->unsigned();
 
             $table->primary(['id_wine', 'id_menu']);
